Increases the rate at which consume_later() can ingest messages.
Originally, the background thread used sleep(3) for 1 second so that the
main thread could reliably acquire the mutex. However, this effectively
limited the maximum frequency to 1 Hz, found in practice to be too low.

Instead, we use nanosleep(2) for a new max frequncy of 100 Hz.

As a nice side effect, the consume_later() tests now run much faster.