Set QoS on a channel *before* calling basic_consume().
It turns out that QoS needs to be set beforehand, otherwise it will
only apply to *new* consumers (of which there are usually none).

This means that the prefetch count was basically never working.

It also resolves the strange behaviour with prefetch counts in the
RabbitMQ management plugin UI.

See also:

- https://github.com/rabbitmq/rabbitmq-management/issues/311 and
- https://www.rabbitmq.com/consumer-prefetch.html

This also makes QoS apply even if `no_ack` is TRUE -- which was likely
another unrelated mistake on my part.