Refactor consumer callbacks to (n)ack *after* execution.
The existing behaviour is basically the same as always having
no_ack=TRUE, in that messages are unequivocally acknowledged. However,
it is sometimes desirable to distinguish processing which fails from
that which succeeds, so callbacks which throw an error will now nack
messages instead -- without requeing them.

Since nack without requeue is only *semantically* (not operationally)
different than acknowledgment, existing workflows are unlikely to be
affected.

Documentation has been updated to reflect this behaviour.

Part of #2.